pyproject.toml
[tool.poetry]
version = "0.1.0"
name = "llm-index"
readme = "readme.md"
description = "rag for llm"
authors = ["elamribadrayour <badrayour.elamri@protonmail.com>"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.11"

pandas = "2.2.0"
db-dtypes = "1.2.0"
fastparquet = "2024.2.0"
azure-storage-blob = "12.19.0"
google-cloud-storage = "2.14.0"
google-cloud-bigquery = "3.17.2"

[tool.poetry.group.ci.dependencies]
ruff = "0.2.1"
mypy = "1.8.0"
pytest = "8.0.0"
black = "24.1.1"
pandas-stubs = "2.1.4.231227"
types-requests = "2.31.0.20240125"
